+++ CHANGELOG.rst
@@ -3,6 +3,13 @@
 Change Log
 ==========
 
+0.0.10
+++++++
+  * Now a namespace package
+  * Bug Fix:
+
+    + When targeting Capture One packages with directories containing thumbs.db or .DS_Store files, these files are simply ignored
+
 0.0.9
 +++++
   * Add the ability to transform between the following package types
@@ -13,5 +20,3 @@ Change Log
 
 Dev
 +++
-
-  * Now a namespace package